Add Karma testing
Unit testing is another important idea. Usually when we want to see if something works, we mess around with it in the console and if we get what we want we are happy. With unit testing, its the same idea, except we write what we would have in the console in an actual file. We run that file with a test-runner, and the test runner tells us if we are passing or failing. This seems kind of dumb at first, but overtime it is really useful. If we make one change to our code, we can instantly see if the unit tests are passing or not. If not, then we know we messed something up.